+++ R/zzz.R
@@ -1,11 +1,5 @@
-#' @importFrom cli style_bold
-#' @importFrom crayon blue yellow green red
 #' @importFrom curl curl curl_download
 #' @importFrom glue glue
-#' @importFrom magick image_read image_read_svg
-#' @importFrom memoise memoise
-#' @importFrom raster extract getData raster crop
-#' @importFrom rsvg rsvg_raw
 #' @importFrom sf st_read
 #' @importFrom utils unzip
 NULL
@@ -15,40 +9,8 @@ NULL
 
 gh_url <- "https://github.com/"
 ghraw_url <- "https://raw.githubusercontent.com"
-
-glue_url <- function(...) glue(..., .sep = "/", .envir = parent.frame(1))
-
 gadm_url <- "https://biogeo.ucdavis.edu/data/gadm3.6/"
-
-lu_url <- "https://www.agr.gc.ca/atlas/data_donnees/lcv/aafcLand_Use/tif/"
-
-# see inSilecoMisc
-
-msgInfo <- function(..., appendLF = TRUE) {
-  txt <- paste(cli::symbol$info, ...)
-  message(blue(txt), appendLF = appendLF)
-  invisible(txt)
-}
-
-
-msgError <- function(...) {
-  txt <- paste(cli::symbol$cross, ...)
-  message(red(txt))
-  invisible(txt)
-}
-
-msgSuccess <- function(...) {
-  txt <- paste(cli::symbol$tick, ...)
-  message(green(txt))
-  invisible(txt)
-}
-
-msgWarning <- function(...) {
-  txt <- paste(cli::symbol$warning, ...)
-  message(yellow(txt))
-  invisible(txt)
-}
-
+glue_url <- function(...) glue(..., .sep = "/", .envir = parent.frame(1))
 
 # see graphicustils